refactor(hook): theme 정보 조회 로직 useThemeInfo 훅으로 추출
- ThemeHero 컴포넌트 내 axios 호출, error/로딩 상태를 useThemeInfo로 분리해주었습니다.
- ThemeHero는 UI만 담당하는 컴포넌트로 역할을 분리해주었습니다.
